Add function close menu when user scrolling
